Ported recent changed to 64 bit platform.
git-svn-id: https://aug.svn.sourceforge.net/svnroot/aug/trunk@89 a3a97f7c-8e16-0410-9893-1d5f7d0f8f6e

+++ example/httpd/state.c
@@ -89,7 +89,7 @@ prepare_(aug_state_t state)
     state->out_.iovec_[0].iov_len = aug_strbuflen(state->out_.header_);
 
     if (next->mar_) {
-        size_t size;
+        unsigned size;
         state->out_.iovec_[1].iov_base
             = (void*)aug_content(next->mar_, &size);
         if (size) {
